# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ORCID Link Service",
        "description": "The *ORCID Link Service* provides an API to enable the linking of a KBase\n user account to an ORCID account. This \"link\" consists of a [Link Record](#user-content-header_type_linkrecord) which \n contains a KBase username, ORCID id, ORCID access token, and a few other fields. This link record allows\n KBase to create tools and services which utilize the ORCID api to view or modify\n certain aspects of a users ORCID profile.\n\nOnce connected, *ORCID Link* enables certain integrations, including:\n\n- syncing your KBase profile from your ORCID profile\n- creating and managing KBase public Narratives within your ORCID profile",
        "termsOfService": "https://www.kbase.us/about/terms-and-conditions-v2/",
        "contact": {
            "name": "KBase, Lawrence Berkeley National Laboratory, DOE",
            "url": "https://www.kbase.us/",
            "email": "engage@kbase.us"
        },
        "license": {
            "name": "The MIT License",
            "url": "https://github.com/kbase/kb_sdk/blob/develop/LICENSE.md"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/orcid/works/{put_code}": {
            "get": {
                "tags": [
                    "works"
                ],
                "summary": "Get Work",
                "description": "Fetch the work record, identified by `put_code`, for the user associated with the KBase auth token provided in the `Authorization` header",
                "operationId": "get_work_orcid_works__put_code__get",
                "parameters": [
                    {
                        "name": "put_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The ORCID `put code` for the work record to fetch",
                            "title": "Put Code"
                        },
                        "description": "The ORCID `put code` for the work record to fetch"
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Link not available for this user"
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "works"
                ],
                "summary": "Delete Work",
                "operationId": "delete_work_orcid_works__put_code__delete",
                "parameters": [
                    {
                        "name": "put_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Put Code"
                        }
                    },
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Work record successfully deleted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orcid/works": {
            "get": {
                "tags": [
                    "works"
                ],
                "summary": "Get Works",
                "description": "Fetch all of the \"work\" records from a user's ORCID account if their KBase account is linked.",
                "operationId": "get_works_orcid_works_get",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ORCIDWorkGroup"
                                    },
                                    "title": "Response 200 Get Works Orcid Works Get"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Link not available for this user"
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "works"
                ],
                "summary": "Save Work",
                "description": "Update a work record; the `work_update` contains the `put code`.",
                "operationId": "save_work_orcid_works_put",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Link not available for this user"
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "works"
                ],
                "summary": "Create Work",
                "operationId": "create_work_orcid_works_post",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "minLength": 32,
                                    "maxLength": 32
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "KBase auth token",
                            "title": "Authorization"
                        },
                        "description": "KBase auth token"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/NewWork"
                                    }
                                ],
                                "description": "The new work record to be added to the ORCID profile.",
                                "title": "New Work"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Work record successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Work"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "KBase auth token absent or invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Input or output data does not comply with the API schema",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string",
                        "maxLength": 25,
                        "minLength": 5,
                        "title": "Code",
                        "description": "A unique code associated with this specific type of error"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5,
                        "title": "Title",
                        "description": "A human-readable title for this error; displayable as an error dialog title"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "A human-readable error message, meant to be displayed to an end user or developer"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ServiceBaseModel"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "title",
                    "message"
                ],
                "title": "ErrorResponse",
                "description": "\n    A generic error object used for all error responses.\n\n    See [the error docs](docs/errors.md) for more information.\n    "
            },
...
            "ORCIDWorkGroup": {
                "properties": {
                    "updatedAt": {
                        "type": "integer",
                        "title": "Updatedat"
                    },
                    "externalIds": {
                        "items": {
                            "$ref": "#/components/schemas/ExternalId"
                        },
                        "type": "array",
                        "title": "Externalids"
                    },
                    "works": {
                        "items": {
                            "$ref": "#/components/schemas/WorkSummary"
                        },
                        "type": "array",
                        "title": "Works"
                    }
                },
                "type": "object",
                "required": [
                    "updatedAt",
                    "externalIds",
                    "works"
                ],
                "title": "ORCIDWorkGroup"
            },
...
            "ServiceBaseModel": {
                "properties": {},
                "type": "object",
                "title": "ServiceBaseModel"
            },
...
        }
    },
...
}